// web/sanity.types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export type SanityImageDimensions = {
  _type: "sanity.imageDimensions";
  height?: number;
  width?: number;
  aspectRatio?: number;
};
// ...
export type PROJECTS_QUERY_RESULT = Array<{
  _id: string;
  title: string | null;
  slug: Slug | null;
  eyebrow: string | null;
  summary: string | null;
  tags: Array<string> | null;
  url: string | null;
  linkLabel: string | null;
  order: number | null;
  image: {
    asset: {
      _id: string;
      url: string | null;
      metadata: {
        lqip: string | null;
        dimensions: SanityImageDimensions | null;
      } | null;
    } | null;
    alt: string | null;
    hotspot: SanityImageHotspot | null;
    crop: SanityImageCrop | null;
  } | null;
}>;